Web-App/app/requirements.txt
Django==6.0.4
gunicorn==23.0.0
whitenoise==6.9.0
asgiref==3.11.1
sqlparse==0.5.5
google-auth==2.49.2
google-auth-oauthlib==1.3.1
google-api-python-client==2.194.0
google-genai==1.73.1
googleapis-common-protos==1.74.0
google-auth-httplib2==0.3.1
httplib2==0.31.2
requests==2.33.1
requests-oauthlib==2.0.0
oauthlib==3.3.1
beautifulsoup4==4.14.3
python-dotenv==1.2.2
httpx==0.28.1
